requirements.txt
psycopg2-binary==2.9.5
pytest==7.2.2
python-dotenv==1.0.0
testing.postgresql==1.3.0